+++ src/app/middleware/exception_handler.py
